Rename reviews.py to Places_reviews.py
