+++ api/v1/views/__init__.py
@@ -8,4 +8,4 @@ from api.v1.views.index import *
 from api.v1.views.cities import *
 from api.v1.views.states import *
 from api.v1.views.users import *
-from api.v1.views.reviews import *
+from api.v1.views.places_reviews import *
